requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.0
pydantic-settings==2.6.1
python-dotenv==1.0.1
httpx==0.27.2
json-repair==0.44.1
SQLAlchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.19
starlette>=0.35.0
bcrypt>=4.0.0
playwright>=1.49.0
pycryptodome>=3.20.0
tos>=2.9.0
Pillow>=10.0.0
twilio>=9.0.0,<10
PyJWT>=2.0.0,<3.0.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-httplib2>=0.2.0
httplib2>=0.22.0
PySocks>=1.7.0
